refactor: remove unused components and rename new
